fix(proxy): move delta to config (#968)
* Move delta to Config

We parameterise Config with a D parameter that represents the delta
between timestamps. The Sub impls then specialise their output to use
this D as the Output.

* Move the delta concern to Requested states

For the case of Created and Found requests we would like to ignore the
deltas and always return the requests. This means that filter can just
be in terms of RequestState and the filter on deltas is moved to the
particular case of Requested.

* Fix naming and tests

Clippy was constantly getting our filter confused with the filter in
Iterator so we introduce the *_by_state suffix.

We also fix the tests with the new code and naming changes.