Allow similar names
This seems more of a hindrance than anything. I feel like it's a rare
occasion when someone writes a name that is different only by 1
character, and when they do its a genuine difference, such as urn and
url.
To be honest, name shadowing seems more dangerous than "similar names"
and this lint doesn't even bother with that.