feat(proxy): waiting room subroutine (#967)
Closes #955 

***
* Start streamin
* Incant the right arcana
Cleaning up the test of getting a query stream and transitioning the
state of the requests within the stream.

* Start attack of the clones
* Attack of code golf
Clean up the implementation of next_clone.
Fix the test so that has found the peer and transitions that request to
cloning.

* Get a clonable immediately
* Move delta to Config
We parameterise Config with a D parameter that represents the delta
between timestamps. The Sub impls then specialise their output to use
this D as the Output.

* Move the delta concern to Requested states
For the case of Created and Found requests we would like to ignore the
deltas and always return the requests. This means that filter can just
be in terms of RequestState and the filter on deltas is moved to the
particular case of Requested.

* Hook up queries and clones stream
* Track found URns
* Fix naming and tests
Clippy was constantly getting our filter confused with the filter in
Iterator so we introduce the *_by_state suffix.

We also fix the tests with the new code and naming changes.

* Remove delta from the stream structs
* Introduce Shared type
This type makes it easy to use the `Into` pattern for something that
needs to be held in a Arc/RwLock.

* Use Into Shared for WaitingRoom
And fix the call sites of run by passing a fresh WaitingRoom as a
parameter.

* Use RadUrl as parameter for WaitingRoom functions
Since we're taking a RadUrn and a PeerId, it would make sense to unify
these and take a RadUrl instead. This makes it easier for the callers
since they will generally be dealing with RadUrls.

* Pass in WaitingRoom at main
* Format gossip tests
* Add From impl for Arc+RwLock for Shared
* Format replication tests
* Create new RequestCommand
We create a separate RequestCommand that gets embedded in Command. This
allows us to write a simple run function that holds the logic for
executing a command and logging the errors.

* Fix and format main
* Fix tests related to waiting room
* Remove repo parameter and use the urn from the request
* Add tests for transitions
* Allow similar names
This seems more of a hindrance than anything. I feel like it's a rare
occasion when someone writes a name that is different only by 1
character, and when they do its a genuine difference, such as urn and
url.
To be honest, name shadowing seems more dangerous than "similar names"
and this lint doesn't even bother with that.

* Reorganise and clean up
Moving the run function on RequestCommand to be a private function in
peer.
Document all the things.
Clippy all the things.

* HOLY SMOKES IT WORKS
* Move request into a sub-module
* Cloned takes the RadUrl from where it was cloned from
* Fix up to docs, comments, etc.
* Document request module
* Internalise the delay and bump down to 50ms
* No need for a second function
* Bump timeout for stream tests
* Split commands into separate functions
We split the request command evaluation into separate functions. We then
follow the idiom found in peer to spawn these as sub-tasks and
communicate success/failure cases.

* Next query can be slightly optimised
We can return a Requested request if it hasn't attempted to clone yet so
that we return it without waiting for the elapsed period.

* Enable logging for coco tests
* Ensure we only create requests with empty paths
When we're cloning we don't need to care about paths in RadUrns. So when
we create a request we use the empty path. Then to simplify doing this
every time for a WaitingRoom we key the HashMap by the Hash of the
RadUrn.

* Wait on requested + clean up
* Clippy
* Log trace
* Want to see if the timeout delay matters in CI
* Core threads?!
* Disable logging on other tests
* Disable logging on other tests
* Disable logging on other tests
* Disable logging on other tests
* Narrow down logging
* Ping the `select!` in the event loop
We noticed that having a branch that always succeeds drives the event
loop forward. We need to investigate why this is needed.

* Turn back time
* Format
* Fix waiting room passing
Co-authored-by: FintanH <fintan.halpenny@gmail.com>