Serde Macro Magic
We can utilise serde's macro magic using the `with` and `bound` macros
to tell serde that we'll use the Milliseconds trait for serialisation.

Note that this enforces to use the time types defined in serde_millis
when we want serialise any of these types. Otherwise, we will have to
hand-roll implementations.