fix(proxy): drain peer channels
The polling of subroutines can starve if the channels are not completely
drained. async/await while being young already is a form of witchcraft
only the elder gods can control. Who knows what else lurks in the depth
of its abyss. Only the brave willl risk the decent into madness.