Rework checkout logic
We set up checking out projects differing by where the project is coming
from.
If it's a owned by the user then we simply clone the repo calling
the remote `rad`.
If it's coming from another peer's branch then we set it up via their
remote. We then have to create the `rad` remote to set that up as our
upstream.

This all culminates in testing that we can clone from a specific peer
and that we don't end up getting changes from another peer that we know
about.