Initialise a Browser based on some ref heuristics
Previously, when initialising a `Browser` we expected the initial branch to be our own default
branch. Now that we're getting into the replicated end of things, this won't always work. When we
first replicate we are only tracking the project, and that means we don't have a copy of the default
branch under our view of the project.

To fix this, we need to look at our peer's views, and in particular, we can use the first maintainer
to initialise the `Browser`.

So the initialisation of a Browser now checks for our own branch first in the storage, if that
doesn't exist then fall back to the first maintainer.