with_browser takes a NamespacedRef
We now have with_browser taking a NamespacedRef which can be obtained by
find_default_branch or get_branch. Which of these two functions is to be
used is dependent on whether we want to ensure which branch we're
starting off with or if it's good enough to get some default branch for
the project.