fix(proxy): peer browsing (#1015)
* Initialise a Browser based on some ref heuristics

Previously, when initialising a `Browser` we expected the initial branch to be our own default
branch. Now that we're getting into the replicated end of things, this won't always work. When we
first replicate we are only tracking the project, and that means we don't have a copy of the default
branch under our view of the project.

To fix this, we need to look at our peer's views, and in particular, we can use the first maintainer
to initialise the `Browser`.

So the initialisation of a Browser now checks for our own branch first in the storage, if that
doesn't exist then fall back to the first maintainer.

* Split out the logic for getting a branch

* Clean up via clippy

* Use NamespacedRef as parameter instead

* with_browser takes a NamespacedRef

We now have with_browser taking a NamespacedRef which can be obtained by
find_default_branch or get_branch. Which of these two functions is to be
used is dependent on whether we want to ensure which branch we're
starting off with or if it's good enough to get some default branch for
the project.

* Bump librad's SHA

* Update after the merge

* Fmt