feat(ui): check if project exists when searching (#1030)
Up until now we assumed the provided urn is not present in the monorepo.
This change alters the behaviour to first check if a project exists
locally - because locally sourced is best source. If it exists locally
the result box offers to navigate directly to it -- also possible with a
hit of enter. Should it not exist a track toggle is presented, which
when pressed will fire a request for the project and show a
notification.

    check for project locally first
    navigate to existing project
    request urn from the network on track toggle
    validate input to be a valid urn

Refs #984

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>