refactor(proxy): Use signer in `coco::State`
We remove the `BoxedSigner` argument from all `coco` functions that act
on `coco::State`. Instead we keep the `BoxedSigner`, which is provided
when creating `coco::State`, around as a field.

The motivation for this is to centralize the signer management in
`coco::State` which is crucial for the general key management operation.