fix(proxy): checkout peer (#1010)
* WIP: Attempt to test the checkout behaviour between 3 peers

* Rework checkout logic

We set up checking out projects differing by where the project is coming
from.
If it's a owned by the user then we simply clone the repo calling
the remote `rad`.
If it's coming from another peer's branch then we set it up via their
remote. We then have to create the `rad` remote to set that up as our
upstream.

This all culminates in testing that we can clone from a specific peer
and that we don't end up getting changes from another peer that we know
about.

* Bump librad

* Bump librad

* Update Cargo.lock

* Update docs on handle field of Remote

* Fix test after the merge

* Reorganise and document

We move the clone functions as methods on Ownership to make things
clearer and compartmentalised.

We also document the checkout verbosely, because it's not very clear
from the actions take but is clearer when described and shown what the
final config looks like.

* Fixes to the UI code

- remote should have been called peerId and is optional
- passing the PeerId was stubbed so we pass down the currentPeerId now
- the endpoint didn't have the full path since it was missing
`/checkout`. It must have been a fluke that it was a match regardless.

* Fix fetch refspecs

It turns out we were using the refspecs and the URLs wrong. We create the
correct refspecs here and clean up the documentation and parameter
passing.

* Formattergit diffgit diffgit diff!

* Whoop(s)

* Move guards to http

Since we need the guard for self peer id in checkout as well, we move
them to the http module so that they can be reused. But they should
eventually be removed once we figure out how to handle this better.

* Cliippeeee

* Pass the peer_id to get_project

* Get out the clippers