Persist TimedOut state (#1059)
We never persisted the TimedOut state when we encountered one, instead
returning it directly as an error.
This PR fixes that by first persisting the TimedOut state during a
transition and then returning it as an error. We can confirm the
persistence in the test cases for when we reach a timeout.