chore(ui): drop old registry code
Continued improvements to use types in the ui code let to more and more
overhead keeping the registry related code in sync which lays barren.
This is a liability and should we be in need to resurrect any of the
valuable work we can always pick and choose from an older state.