refactor(proxy): Introduce Context(Un)sealed
We make the `Context` type into an enum so that the API is forced to
handle the case where the keystore has not been unsealed yet and
consequently the coco state and peer are not available.

Currently, the API is always intialized with an unsealed context. Making
dynamic changes will be introduced in a follow-up.