fix(proxy): avoid double discovery on startup
As the initial value is set when the watch channel is created any
subsequent calls will produce another message. So far the caching of the
seed to avoid continues sending of discovery updates was done in the
main binary. This led to besides the initial value another one was sent
on the first iteration of that task. Leaving the Peer in an offline
state after one of them got disconnected.