feature(proxy): list remotes (#1081)
We refactor tracked so that it returns the most information as possible.
There are three cases in particular:
* Searching -- peer is tracked but no information is available for them
yet
* Maintainer -- peer is a maintainer of the project
* Tracking -- peer is a contributor to the project

***
* Rework semantics
* Fmt
* Make endpoints consistnet
* Fix path for peers endpoint
* Change to peer abstraction
* Everybody loves Functors
* Moar docs
* Fix typo
* Refactoring
Move Peer into its own module
Refactor revisions to use Peer

* Replace filter_maps
* Rework replication to be a type witness
We make the status field of Peer generic so that we can allow it to be a
type witness. Status is the general replication status, where Replicated
is the interesting case. This allows us to talk about only Replicated
Peers.

* Fix tests after refactoring
* Let me see you documents, please
* User project::Peer instead of generic Into
* TODOcument
* Sorry fake peers
Tracking fake peers was always a bad idea, and I'm learning the error of
my ways now. We'll always get an error whenever we try to access
signed_refs using a different PeerId.

* Cleanup clippy pragmas and imports
* Clippy + Formatter - my two archnemeses
* Move docs
Co-authored-by: Alexander Simmerl <a.simmerl@gmail.com>
Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>