fix(ui): show correct projects for users
The underlying issue was an automatic preloading on store initialisation
which fetched the owners projects. It would override the list of fetched
projects for the accessed user profile. To avoid similar subtle bugs
when building up data to satisfy the screen, all stores for the
UserProfile screen and its fragments are isolated into their own ts file
with a streamlined public api.

Fixes #1100