fix(proxy): partial project lists (#1129)
* Allow similar names

* Drafting up partial failure

When listing projects we don't want a single project to be a show-stopper
for the rest of the projects. So we capture the failures, logging them
as warnings, and create a failure list.

* Differentiate a project with or without Stats

We parameterise Project by its stats. There are two cases, we either are
missing Stats but have the rest of the metadata, or we have stats _and_
the metadata. These are called Partial and Full respectively.

* Remove TODO

* Go Clippers!