Tidy up errors
We embed validation::Error into create::Error so that it be resurfaced
through that enum. This allows us to re-validate upon setup, Just In
Case.

We also allow empty directories to become a new working copy, but rule
out directories that already have contents, or paths that lead to files.