fix(proxy): validate creation paths (#1134)
* Validate creation paths

We don't want to create a radicle project and its working copy if it fails
validation criteria, since we can't rollback that creation. In this change
we make sure that a path is validated before finishing the attempt to
create the radicle project and its working copy.

* LA Clippy

* Fix docs

* Format

* What's up docs

* Return 400 on NotARepo

* An empty directory is still valid for creating a new project

* Tidy up errors

We embed validation::Error into create::Error so that it be resurfaced
through that enum. This allows us to re-validate upon setup, Just In
Case.

We also allow empty directories to become a new working copy, but rule
out directories that already have contents, or paths that lead to files.