Refactor validation of project creation
We want to validate some criteria when a project is being created for an
existing working copy. This involves checking the repository for the
default branch and an existing rad remote.

We continue on with the path validation, but also add these validation
checks in as well. To make this more digestable, we move these
validations into their own module, where the create module simply calls
out to this functionality.