fix(proxy): extensive validation (#1190)
* Refactor validation of project creation

We want to validate some criteria when a project is being created for an
existing working copy. This involves checking the repository for the
default branch and an existing rad remote.

We continue on with the path validation, but also add these validation
checks in as well. To make this more digestable, we move these
validations into their own module, where the create module simply calls
out to this functionality.

* Remove invalid tests

These tests were testing some unsavoury behaviour. Our behaviour before
was to delete the remote and create a fresh one. This could lead to
references under that remote being deleted.

Instead, we now error while validating, and the user can decide how to
correct the state of their repository.

* Docs and format

* Simplify Create and fix docs

* Fix le docs

* Fix up tests modules and http error conversion

* Missing "remote" in sentence

Co-authored-by: Alexander Simmerl <a.simmerl@gmail.com>

* Fix up documentaton

Co-authored-by: Alexander Simmerl <a.simmerl@gmail.com>