refactor(ui): Convert native/ipc to TypeScript
We convert `native/ipc.js` to TypeScript. This enables us to use it from
TypeScript files (not only Svelte files) in `./ui`. The reason why we
need to convert the file to TypeScript is that rollup and typescript
don’t seem able to handle JS files and spit out a warning. In general it
also seems to be a good idea to move everything to TypeScript.

One result of this change is that we can remove everything Babel
related.