Build custom Wallet QRCode modal
Note: closing the modal now doesn't disconnect the attempt as the API
provided to build a custom modal doesn't provide an event for that and
we don't have that either. TODO.

Note: Trying to follow the Rimble design approach here as an experiment.
It might be smart to follow their convention to build a seamless
experience for crypto users. It looks great too.