Update include file on fetch and applied
When we fetch or when have applied a change we know that a peer that we
tracked but not received changes for yet should now exist in our repo.
This is the opportune time to update the include file for this project
and add the peer's handle as a remote.