Don't set wallet as connecting until initialize
Now, as one clicks on "connect wallet" and gets the QRCode modal, when
one dismisses that modal and gets back to the starting screen, the
"connect wallet" button is now just spinning, as the call to
"wallet.connect()" has set the wallet's state to "Connecting". To avoid
this, we only set the state to connecting once the user has approved the
connection and we must initialize all things. This makes sense from a
practical perspective as well: we are not in fact connecting until
there's green light from the other side, but just awaiting for the user
to approve the connection. Anyway, for the UX this fixes the problem and
does not cause any quirk.