feat(ui): Introduce structure error logging
This change introduce structured error logging and applies it in a
couple of places.

The design of the `Error` interface is very similar to Rust errors and
arises from the following needs:
* A stable identifier for errors (`code`)
* A message we can display to the user (`message`)
* Structured details that provide more information about the error (`details`)
* A stack trace if available (`stack`)
* Layering of error chains (`source`)

I’ve decided not to make `code` an enum. Having a centralized place that
lists all errors does not help us with debugging. When debugging we need
to look at the place where the error originated which can easily be
found be grepping for the error code. In addition such a centralized
place would make it more cumbersome to add new errors since the list of
error would always need to be updated.

Errors are shown to the user as notifications (`error.show`). The error
message is displayed to the user and the full error can be copied to the
clipboard. In addition we also log the error to the console which allows
us to debug it during development.

As a follow-up to this PR we will apply the error logging and showing in
more places.