feat(ui): reduce number of retries after proxy restart
With #1153 the proxy restarts much quicker after the keystore is
unsealed or created. This allows us to reduce the number of retries and
fail earlier if the proxy does not start properly. (We now wait 5
seconds).

As part of this change we also adjust the signature of `withRetry`. Now
the caller is able to control all timing related parameters.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>