feat(ui): introduce structured error logging (#1221)
This change introduce structured error logging and applies it in a
couple of places.

The design of the `Error` interface is very similar to Rust errors and
arises from the following needs:
* A stable identifier for errors (`code`)
* A message we can display to the user (`message`)
* Structured details that provide more information about the error (`details`)
* A stack trace if available (`stack`)
* Layering of error chains (`source`)

Errors are shown to the user as notifications (`error.show`). The error
message is displayed to the user and the full error can be copied to the
clipboard. In addition we also log the error to the console which allows
us to debug it during development.

As a follow-up to this PR we will apply the error logging and showing in
more places.

Co-authored-by: Merle Breitkreuz <merle@monadic.xyz>
Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>