refactor(coco): Simplify `CreateRequest` control input (#1288)
`input::Control::CreateRequest` now always holds a response sender
instead of it being optional. The `None` case was only used in the tests
and the variant is now aligned with the other `input::Control` variants.

This also allows us to simplify the tests. We remove any assertions on
the command output being empty,