fix(proxy): only save most recent logs from proxy
We change how the proxy process is spawned. Instead of saving all output
of the process in a buffer we only save the 200 most recent lines. In
addition we forward all of the proxy’s output to the main process
output. This allows us to see the proxy logs when starting the app from
the terminal.

To implement this we introduce the `ProxyProcessManager` which
encapsulates this behavior. The logic for line buffering the output of
the process suffers from duplication. I’ve experimented with some
abstractions but they were all to complicated and did not add enough
value.

Fixes #1359