Fix fatal error on withdrawing
On withdrawing, the app would crash with a console error related to a
remote store and calling `stop` whereas it's not a function. That
happened in this piece changed here, where we were subscribing on the
pool's remote store, which would be left when the modal was closed.

Now we are just reactively listening to it, instead of being subscribed
to it. Seems to have fixed the issue and the solution doesn't seem bad.