fix(ci): put timeout on proxy tests
We’re using the `--ensure-time` option with `cargo test` to notify us
when a test takes longer than the specified time. However, this does not
abort the test suite and thus does not address #1378. To actually abort
the test suite we are using the `timeout` command.

In addition we have the following changes
* We remove the `timeout` command from the lint step. This was committed
  by me accidently in #1379. We have never experienced issues with
  clippy timing out.
* We’re replacing the `--ensure-time` option with `--report-time`. We
  want to know how long tests take but not fail them since we have no
  expectations for the time they should take.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>