Offer ergonomic in-between Pool Contract API
The underlying Contracts API is far from ergonomic for fair reasons:
Keep the contracts as simple as possible and its API's area as small as
sensible. Here we wrap the underlying contract in a class that offers
all the ergonomics we wish, separating the logic of how to translate a
sensible use case (TopUp) to how the contract interfaces it
(updateSender(topUp, 0, UNCHANGED, [], []). This also helps with not
having to deal with both this translation and with the pipeline that
such use cases trigger in the app.