Make SidebarLayout better responsive
The Funding > Outgoing > Support component packs a good number of
elements horizontally, which fall on top of each other in smaller screen
sizes. Setting a min-width bigger than the global --content-min-width
value to the Support component container does not work well neither is
it optimal since it would only increase the width of that element alone
but now of the whole side-bar layout, ending up with the header with a
shorter width than the content below and the whole thing looking poor.

This solution, discussed and agreed on with Brandon, results in the
SidebarLayout being responsiveness to whatever children's width and
always - at least - take the width from the side bar until the right end
of the viewport.