s/erc20/dai
The idea was to decouple the usage of a ERC20 token from the token in
use itself, DAI, but since the token we use has direct implications in
the UI when showing the token used, we better be explicit in the code as
well. In the future, when we might want to support different tokens, we
can then better model this to be polymorphic over the token employed.