feat(ui): Funding Pool v0
The v0 of the Funding Pool is in alpha. The feature is only available
behind the isExperimental and fundingFeature flags. It operates using a
local ethereum node. All relevant documentation can be found in
FUNDING.md.

This is the result of a joined effort of multiple people. A special
thanks to @geigerzaehler, @CodeSandwich, and @brandonhaslegs.