Handle old rad remote
If the URL could not be parsed then it's likely that it was an old rad
remote. We emit a warning, rename the old one to `temp_rad` and create a
fresh `rad` remote. This allows users to remove the `temp_rad`
themselves and handle any of their old references if necessary.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>