Gather NotificationFaucet and TransactionCenter
We gather these two elements in a new 'Footer' component, which
allows us to display them side by side without overlap and without
laborious and highly-dependant calculations. The caveat of this solution
is that the footer needs a default z-index, meaning that the
TransactionCenter is too high in this plane, thus the need to hide it
when a modal is showing.