test: commit replication (#1589)
* Automate remaining networking QA steps
* Use pasteInto to speed up specs
* Run integration tests in --release mode to speed them up
* Rename storagePath to radHome
* Don't block the plugin thread
* Fix race condition in remote management modal

The remote list in the modal sometimes wouldn't update correctly showing
a newly added remote in a "still looking" state forever. This happened
consistently on CI because of subtle timing differences.

The cause of this bug was that clicking the "Add" remote button would
trigger a backend request to get the updated list, at this moment the
list would still contain the remote in a "still looking" state. Meanwhile
the backend would find the remote and fire a refresh event that the UI
would pick up and update the store to show the found remote, this all
happens before the first request completes. At this moment the remote
list shows the correct state, but now the first request finishes and
re-updates the list with the old value showing the remote again in a
"still looking" state.

To prevent the race condition we additionally add the locking mechanism
to the fetchPeers function.

Closes #1529

Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>