Rethink solution and test the two possible cases
In order to stub the fetching of the users' git default branch we had to
switch to use the RendererMessage flow. We do that and test both the
case where that property is obtained and when it fails to be obtained,
resulting in 'trunk' and 'main' being set as the default branches,
respectively.