feat(test): store test node logs in file
We store the output of nodes in tests in files instead of printing it to
stdout. This declutters the test output and makes it easier to inspect
the logs, especially on CI. On CI we upload the logs as artifacts.

To make it easy to find the logs we adjust the function that creates
temporary directories so that it uses the test name instead of a UUID.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>