fix(proxy): announce heads and tags with correct path
We properly announce changed refs. Instead of announcing the urn
`rad:git:<project-id>:foo` if the `foo` branch changed we announce
`rad:git:<project-id>:refs/heads/foo`. This is what `radicle-link`
expects. In addition we also announce tags.

I’ve also changed the visibility of methods in the module that are
unused externally.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>