refactor(ui): Download twemoji assets directly
Instead of using our own `twemoji-svg-assets` package we download the
assets directly as part of the `postinstall` scripts.

This has two advantages: First, we don’t need to maintain a separate
package. Second, this removes a blocker of the [upgrade to Yarn v2][yv2]. Our
`twemoji-svg-asstest` package copied the assets as part of *its*
[post-install script][pis]. This only works under certain assumptions
about the directory layout which don’t hold for Yarn v2.

[pis]: https://github.com/radicle-dev/twemoji-svg-assets/blob/master/package.json#L10
[yv2]: https://github.com/radicle-dev/radicle-upstream/issues/1687

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>