wip: Check attestation status
Need to add @typechain/etheres-v5@6.0.0 to the dependencies in order to
access the events emited by the Claims contract and be able to actually
look up claims - if any - by the connected eth account.

I am also seeing the funding screen flickering.

Finally, this state-dependency chain is turning to hell. We should
probably have a global 'EthereumStatus' type that encapsulates all the
possible states and holds the values each state entails.

Signed-off-by: Nuno Alexandre <hi@nunoalexandre.com>