raidcle-upstream upgrades involving radicle-link's `next` changes
There were some major changes in the radicle-link project that required
radicle-upstream to update to in order for it to move forward on upcoming
features. This patch introduces those changes and a short list of those changes
are outlined below.

The [Identity
specification](https://github.com/radicle-dev/radicle-link/blob/a8f79e4b557030c07e64f7cfb299c000c754c751/docs/spec/sections/002-identities/index.md)
was introduced and implemented. This replaces the Entity code, using Project,
Person, and LocalIdentity. This is a breaking change that means the Radicle IDs
for all projects and persons are different to the previous Entity Radicle IDs.

The local transport became safer and we can use an API supplied by radicle-link
for safely interacting with working copies.

There was also an overhaul of the networking stack in radicle-link. This
required some of the protocol messages to be removed and/or changed. In turn
radicle-upstream had to adapt to these changes. A bonus of this part of the
upgrade was that the project could bump the version of the tokio dependency to
the latest major version >1.0.

New features were reliant on radicle-link's latest changes. Combining this with
the long life of this branch, some other changes are included in this patch. We
list these changes here as well.

Profile support was added. This means that a person can have multiple
subdirectories under their XDG_* directories. These are keyed by a UUID and
allows the person to switch to different profiles manually.

The first step in Ethereum attestations was added. This allows a radicle-link Person
document to be extended with an Eth address that was claimed on the blockchain.

Signed-off-by: Alexander Simmerl <a.simmerl@gmail.com>
Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
Signed-off-by: Igor Zuk <igor.zuk@protonmail.com>
Signed-off-by: Merle Breitkreuz <merle@monadic.xyz>
Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>
Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>