refactor(proxy): fair select on subroutines
We replace the custom `Future` implementation for `Subroutines` with a
simple `run()` function that uses `futures::select`. One important
consequence is that now `Subroutines::inputs` and
`Subroutines::pending_tasks` are polled fairly. Before `Spending_tasks`
was preferred and we tried to drain it which could lead to a delay in
the handling of `inputs`. (This could be the cause for #1378.)

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>