feat(test): reusable `nodeManager.exec` command
We generalize the `nodeManager.createCommit` command to
`nodeManager.exec` to support other commands in the merge request tests.

This change required us to persist the credential helper setting. We
accomplish this by writing the setting to `.gitconfig` in the node’s
home directory. We use the same mechanism to write user data got
`.gitconfig` which eliminates the need to specify those explicitly every
time.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>