move coco::control
Moving the control module over to api since it wasn't used in coco
(apart from one set of tests easily refactored).

The other changes involved scoping code to only be included as the test
target and updating the imports.

One issue was that since control needs reflike and refpsec_pattern it
needs to include radicle-git-ext.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>