fix(ui): replace `path.active`
We replace the incorrectly used `patch.active` function. Whenever this
function was used it was called with a _rendered_ path like
`/projects/rad:git:123` instead of a pattern. The implementation however
interpreted that path as a pattern and passed it to `regexparam`.

This approach worked as long as no `?` was contained in the path passed
to `path.active`. In my case I added a horizontal menu item where the
`href` field looked like `/foo/bar?filter=true`. This `href` field is
used as the `path` argument and the `?` is treated as a special
character. This resulted in `path.active(href, href)` returning false.

The purpose of detecting whether a certain path is active can be more
easily achieved by comparing the rendered path with `$location`. For the
“loose” match mode supported by `path.active` we use
`$location.startsWith`.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>