chore: update cypress to v7
Since the crashes on Cypress v7 have been fixed (#1752) we can update
it. I’ve revied the [changelog][1]. The only change affecting us is that
uncaught exceptions will result in test failures. This caused issues
when we reset the proxy at the start of a test while the app from the
previous test was still loaded in the window. We fix this by unloading
the app before we reset the proxy.

Fixes #1752

[1]: https://docs.cypress.io/guides/references/changelog#7-0-0

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>